+++ core/src/main/java/feign/interceptor/MethodInterceptors.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package feign;
+package feign.interceptor;
 
 import java.util.ArrayList;
 import java.util.Collections;
